Copy the user's GPU frequency table and voltage configuration to the new dtb
Users like to use the Konabess app to adjust the GPU frequency
table and voltage. This is done by modifying the dtb. If we
directly replace the user's existing dtb with a new one, the user's
configuration of the GPU frequency table and voltage will be lost.

So copy them to the new dtb during installation to preserve them.