AK3: Try to repair the vendor_boot partition that was corrupted by an old version of magiskboot
magiskboot versions before Magisk v28.0 will not process the
vendor_ramdisk of the vendor_boot partition normally. magiskboot
versions after Magisk v28.0 can extract all vendor_ramdisks through
the vendor_ramdisk table.

When the old version of magiskboot repackages the vendor_boot image,
it only updates the ramdisk size information in the vendor_boot
header, but does not update the ramdisk size information in the
vendor_ramdisk table. When we use the new version of magiskboot to
unpack the damaged image, we often get the "Splitting image failed"
error, or the magiskboot process is stuck.

Therefore, let's fix the vendor_boot image. Make the size of the
only vendor_ramdisk stored in the vendor_ramdisk table consistent
with the sum of the sizes of all vendor_ramdisks stored in the
header.

For the vendor_boot image with multiple vendor_ramdisks, we will
not fix it. The best solution is to dirty flash the rom.

The precompiled vendor_boot_fix in tools is compiled using
Frankenstein ndk.