AK3: Uninstall KernelSU LKM which is incompatible with GKI
Now almost all AOSP roms of marble are using the OSS kernel of the
LineageOS team, and their kernel is no longer the standard GKI.
Therefore, KernelSU LKM for android12-5.10 GKI cannot work with
the OSS kernel.

For this situation, some rom maintainers provide a boot image
patched with KernelSU LKM for the OSS kernel. Since Melt Kernel is
still GKI, KernelSU LKM for the OSS kernel is not compatible with
Melt Kernel.

Therefore, we had to uninstall the KernelSU LKM which was
incompatible with GKI.

We determine whether it is compatible with the android12-5.10 GKI
by identifying the clang version contained in `kernelsu.ko`.