AK3: Refuse to install on latest LineageOS rom
On the latest LineageOS rom, the developers completely rewrote the
xiaomi touch driver and rewrote the user space interface.

https://github.com/cupid-development/android_kernel_xiaomi_sm8450/commit/a6cf929530dc02f72b1921cd94888b92f61a6a1b

This makes the behavior of ioctl of /dev/xiaomi-touch completely
different from the original one, which in turn causes the single-tap
gesture and double-tap gesture switches to be completely ineffective.

I refused to continue playing cat and mouse with them.